+++ compiler/semantics/tables/classtable.h
@@ -83,7 +83,7 @@ public:
     static bool isHaveAccessToField(const string& requesterClass, const string& requestedClass, const string& requestedField);
 
     static int addIntToConstTable(const string& className, int val);
-    static int addFloatToConstTable(const string& className, double val);
+    static int addFloatToConstTable(const string& className, float val);
     static int addStringToConstTable(const string& className, const string& str);
     static int addClassToConstTable(const string& className, const string& addingClassName);
     static int addMethodRefToConstTable(const string& className, const string& addingClassName, const string& method, const vector<DataType> &params,const DataType &returnType);
